Don't set navStore.partnerOrg on petition fetch
This seems to have been meant to show the petition org in the nav
when a petition was fetched, but this doesn't make sense to me
because there are many times a petition is fetched and nothing
should be shown in the nav.

This also wasn't working at all because the api returns
organization_logo_image_url, and we were looking for logo_image_url.